3.0.0.17 - fix #3163
Fix modal close issue.
Resolves https://github.com/nilsteampassnet/TeamPass/issues/3163

Problem appears to be that we are hiding the modal while it is still
transitioning, so we need to wait on the shown.bs.modal event and also
hide the modal there.
This only happens if the encryption occurs quickly (e.g. for a new user
being enabled from LDAP).